refactor: add deno_file op crate (#10019)
Also enables WPT for FileReader.
