+++ runtime/lib.rs
@@ -5,6 +5,7 @@
 pub use deno_console;
 pub use deno_crypto;
 pub use deno_fetch;
+pub use deno_file;
 pub use deno_url;
 pub use deno_web;
 pub use deno_webgpu;
